Renamed reset-password.ts to reset-password-request.ts
