+++ UT4MasterServer.Web/src/types/reset-password-request.ts
